lots and lots again. make Shader into instance, don't import Drawable into Instance, make services work with Service<RenderingService>, add a transform option to meshes rather than doing it during the loop, expose MeshPart:load_from_obj_file properly, fix a bunch of issues related to headers being imported elsewhere (part of making Shader into instance), expose more IOService methods to Luau, make ReflectionDescriptor have the NotCreatable flag by default and remove it on add_constructor, add std::filesystem::path support to arguments

+++ src/engine/instances/services/RenderingService.h
@@ -24,7 +24,7 @@ namespace Nyanners::Services {
 
         float deltaTime = 0.0f;
         float frameTime = 0.0f;
-        int fps = 0.0f;
+        double fps = 0;
 
         RenderingService() : Instance("RenderingService") {};
